Atualizei os models para ficar mais legiveis ao usar notações do lombok. Após isso fiz pequenas mudanças de consistencia nos repositorios. Após isso fiz pequenas mudanças de consistencia nos services devido a pequenos conflitos com os novos repositorios. Por fim pequenas no mudanças nos controladores devido as mudanças nos services. Estou começando a criar os documents e repositorios para usar elasticsearch

+++ backend/src/main/java/com/IgorAlan/jobportal/repository/JobSeekerApplyRepository.java
@@ -10,7 +10,7 @@ import com.IgorAlan.jobportal.models.JobSeekerProfile;
 import java.util.List;
 
 @Repository
-public interface JobSeekerApplyRepository extends JpaRepository<JobSeekerApply, Integer> {
+public interface JobSeekerApplyRepository extends JpaRepository<JobSeekerApply, Long> {
 
     List<JobSeekerApply> findByUserId(JobSeekerProfile userId);
