+++ backend/src/main/java/com/IgorAlan/jobportal/repository/JobSeekerProfileRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.IgorAlan.jobportal.models.JobSeekerProfile;
 
-public interface JobSeekerProfileRepository extends JpaRepository<JobSeekerProfile, Integer> {
+public interface JobSeekerProfileRepository extends JpaRepository<JobSeekerProfile, Long> {
 
     Optional<JobSeekerProfile> findById(Long id);
 }
