+++ backend/src/main/java/com/IgorAlan/jobportal/repository/RecruiterProfileRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.IgorAlan.jobportal.models.RecruiterProfile;
 
-public interface RecruiterProfileRepository extends JpaRepository<RecruiterProfile, Integer> {
+public interface RecruiterProfileRepository extends JpaRepository<RecruiterProfile, Long> {
 
     Optional<RecruiterProfile> findById(Long id);
 }
